requirements.txt
Django==1.7
argparse==1.2.1
chardet==2.2.1
diff-match-patch==20121119
django-appconf==0.6
django-import-export==0.2.3
nltk==3.0.0
psycopg2==2.5.4
requests==2.4.1
six==1.8.0
tablib==0.10.0
wsgiref==0.1.2
